\name{meta_data}
\alias{meta_Data}
\title{
  Calculate the dataset meta data for use by the Rattle V6 UI.
}
\description{

   The meta data describes the characteristics of the dataset.

}
\usage{
meta_data(df)
}
\arguments{

  \item{df}{A data frame.}
}
\details{

  A function to provide the dataset summary as JSON which can then be
  parsed by Rattle V6 as the dataset summary from which Rattle gets all
  of it's meta data.

}
\keyword{metadata}
